+++ server/src/internalClusterTest/java/org/opensearch/remotestore/RemoteStoreRestoreIT.java
@@ -14,7 +14,10 @@ import org.opensearch.action.index.IndexResponse;
 import org.opensearch.action.support.PlainActionFuture;
 import org.opensearch.cluster.health.ClusterHealthStatus;
 import org.opensearch.common.settings.Settings;
+import org.opensearch.common.unit.TimeValue;
+import org.opensearch.core.common.unit.ByteSizeUnit;
 import org.opensearch.plugins.Plugin;
+import org.opensearch.repositories.RepositoriesService;
 import org.opensearch.test.InternalTestCluster;
 import org.opensearch.test.OpenSearchIntegTestCase;
 import org.opensearch.test.transport.MockTransportService;
@@ -26,9 +29,11 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
 
 import static org.opensearch.test.hamcrest.OpenSearchAssertions.assertAcked;
 import static org.opensearch.test.hamcrest.OpenSearchAssertions.assertHitCount;
+import static org.hamcrest.Matchers.greaterThan;
 
 @OpenSearchIntegTestCase.ClusterScope(scope = OpenSearchIntegTestCase.Scope.SUITE, numDataNodes = 0)
 public class RemoteStoreRestoreIT extends RemoteStoreBaseIntegTestCase {
@@ -448,6 +453,42 @@ public class RemoteStoreRestoreIT extends RemoteStoreBaseIntegTestCase {
      */
     public void testRTSRestoreDataOnlyInTranslog() throws IOException {
         testRestoreFlow(0, true, randomIntBetween(1, 5));
+    }
+
+    public void testRateLimitedRemoteDownloads() throws Exception {
+        assertAcked(
+            client().admin()
+                .cluster()
+                .preparePutRepository(REPOSITORY_NAME)
+                .setType("fs")
+                .setSettings(
+                    Settings.builder()
+                        .put("location", randomRepoPath())
+                        .put("compress", randomBoolean())
+                        .put("max_remote_download_bytes_per_sec", "2kb")
+                        .put("chunk_size", 200, ByteSizeUnit.BYTES)
+
+                )
+        );
+        int shardCount = randomIntBetween(1, 3);
+        prepareCluster(0, 3, INDEX_NAME, 0, shardCount);
+        Map<String, Long> indexStats = indexData(5, false, INDEX_NAME);
+        assertEquals(shardCount, getNumShards(INDEX_NAME).totalNumShards);
+        internalCluster().stopRandomNode(InternalTestCluster.nameFilter(primaryNodeName(INDEX_NAME)));
+        ensureRed(INDEX_NAME);
+        restore(INDEX_NAME);
+        assertBusy(() -> {
+            long downloadPauseTime = 0L;
+            for (RepositoriesService repositoriesService : internalCluster().getDataNodeInstances(RepositoriesService.class)) {
+                downloadPauseTime += repositoriesService.repository(REPOSITORY_NAME).getRemoteDownloadThrottleTimeInNanos();
+            }
+            assertThat(downloadPauseTime, greaterThan(TimeValue.timeValueSeconds(randomIntBetween(5, 10)).nanos()));
+        }, 30, TimeUnit.SECONDS);
+        ensureGreen(INDEX_NAME);
+        // This is required to get updated number from already active shards which were not restored
+        assertEquals(shardCount, getNumShards(INDEX_NAME).totalNumShards);
+        assertEquals(0, getNumShards(INDEX_NAME).numReplicas);
+        verifyRestoredData(indexStats, INDEX_NAME);
     }
 
     // TODO: Restore flow - index aliases